requirements.txt
Appium-Python-Client
pytest==7.1.2
pytest_bdd==5.0.0
requests
sauceclient>=0.2.1
testobject

